+++ workers/tests/integrationtests/api_response_schema.py
@@ -1,42 +1,34 @@
-from typing import Optional
+"""This module defines schemata for API responses.
+
+Every API response must adhere to the general schema.
+Additionally API responses need to adhere to one additional schema depending on which
+endpoint is used.
+"""
 
 from schema import Optional as Opt
 from schema import Or, Schema
 
 
-def generate_properties_template(iterator: Optional[str] = None) -> dict:
-    properties = {
+def get_indicator_properties_template():
+    return {
         "metadata.name": str,
         "metadata.description": str,
         "layer.name": str,
         "layer.description": str,
         "result.timestamp_oqt": str,
-        "result.timestamp_osm": Or(str, None),
+        "result.timestamp_osm": Or(str),
         "result.value": Or(float, None),
         "result.label": str,
         "result.description": str,
-        "result.svg": str,
-    }
-
-    if iterator is None:
-        return properties.copy()
-    else:
-        return {
-            "indicators." + str(iterator) + "." + k: v for k, v in properties.items()
-        }
-
-
-def get_report_properties_template() -> dict:
-    return {
-        "metadata.name": str,
-        "metadata.description": str,
-        "result.value": float,
-        "result.label": str,
-        "result.description": str,
+        Opt("result.svg"): str,
     }
 
 
-def get_response_schema() -> Schema:
+def get_general_schema() -> Schema:
+    """General response schema.
+
+    Every response should have this schema
+    """
     return Schema(
         {
             "apiVersion": str,
@@ -49,19 +41,56 @@ def get_response_schema() -> Schema:
     )
 
 
-def get_feature_schema(iterator: Optional[str] = None):
+def get_result_schema() -> Schema:
+    """Response schema for all endpoints except of 'indicator' and 'report' endpoints"""
+    return Schema({"result": list}, ignore_extra_keys=True)
+
+
+def get_featurecollection_schema() -> Schema:
+    """Response schema for responses of type FeatureCollection"""
+    return Schema(
+        {"type": "FeatureCollection", "features": list}, ignore_extra_keys=True
+    )
+
+
+def get_indicator_feature_schema() -> Schema:
+    properties = get_indicator_properties_template()
     return Schema(
         {
             "type": "Feature",
             "geometry": dict,
             Opt("id"): Or(str, int),
-            "properties": generate_properties_template(iterator),
+            "properties": properties,
         },
         ignore_extra_keys=True,
     )
 
 
-def get_featurecollection_schema():
-    return Schema(
-        {"type": "FeatureCollection", "features": list}, ignore_extra_keys=True
+def get_report_feature_schema(number_of_indicators: int) -> Schema:
+    properties_template = get_indicator_properties_template()
+    properties = {}
+    for i in range(number_of_indicators):
+        for k, v in properties_template.items():
+            prefix = "indicators." + str(i) + "."
+            if isinstance(k, Schema):
+                k._prepend_schema_name(prefix)
+            else:
+                k = prefix + k
+            properties.update({k: v})
+    schema = Schema(
+        {
+            "type": "Feature",
+            "geometry": dict,
+            Opt("id"): Or(str, int),
+            "properties": {
+                "report.metadata.name": str,
+                "report.metadata.description": str,
+                "report.result.value": float,
+                "report.result.label": str,
+                "report.result.description": str,
+                **properties,
+            },
+        },
+        ignore_extra_keys=True,
     )
+    return schema